Removed duplicated code.
